Move test into data_set_manager
